TEST: Split complex test case into two files
