+++ integration_tests/CMakeLists.txt
@@ -343,7 +343,8 @@ RUN(NAME bindc_06   LABELS llvm c
         EXTRAFILES bindc_06b.c)
 RUN(NAME test_generics_01    LABELS cpython llvm c)
 RUN(NAME test_cmath          LABELS cpython llvm c)
-RUN(NAME test_complex        LABELS cpython llvm c)
+RUN(NAME test_complex_01        LABELS cpython llvm c)
+RUN(NAME test_complex_02        LABELS cpython llvm c)
 RUN(NAME test_max_min        LABELS cpython llvm c)
 RUN(NAME test_global         LABELS cpython llvm c)
 RUN(NAME test_global_decl    LABELS cpython llvm c)
