--- tests/reference/asr-test_complex-70f026c.json
@@ -1,13 +1,0 @@
-{
-    "basename": "asr-test_complex-70f026c",
-    "cmd": "lpython --show-asr --no-color {infile} -o {outfile}",
-    "infile": "tests/../integration_tests/test_complex.py",
-    "infile_hash": "0c9a9dbec41e0ad2e587ebdfc9a19e0fdb4a547747f25ac008f31175",
-    "outfile": null,
-    "outfile_hash": null,
-    "stdout": "asr-test_complex-70f026c.stdout",
-    "stdout_hash": "7e1e9fcdc73c6fcf6156e3b19ffc2f84af2e73490a7b963dad70425b",
-    "stderr": null,
-    "stderr_hash": null,
-    "returncode": 0
-}
