+++ tests/reference/asr-test_complex_01-a6def58.json
@@ -1,0 +1,13 @@
+{
+    "basename": "asr-test_complex_01-a6def58",
+    "cmd": "lpython --show-asr --no-color {infile} -o {outfile}",
+    "infile": "tests/../integration_tests/test_complex_01.py",
+    "infile_hash": "073cdd5f46b6ff3c9d5e7e79950d1a4fc46ee760ab04ce8a303afdad",
+    "outfile": null,
+    "outfile_hash": null,
+    "stdout": "asr-test_complex_01-a6def58.stdout",
+    "stdout_hash": "bb3e83880f2c7ddb82cbd1d84bcd7ae99e8ffde68285534adc5146c7",
+    "stderr": null,
+    "stderr_hash": null,
+    "returncode": 0
+}
