+++ tests/reference/asr-test_complex_02-782ba2d.json
@@ -1,0 +1,13 @@
+{
+    "basename": "asr-test_complex_02-782ba2d",
+    "cmd": "lpython --show-asr --no-color {infile} -o {outfile}",
+    "infile": "tests/../integration_tests/test_complex_02.py",
+    "infile_hash": "f89daec2225e1da65f791cac35d151886e4889a3793489271e10fad4",
+    "outfile": null,
+    "outfile_hash": null,
+    "stdout": "asr-test_complex_02-782ba2d.stdout",
+    "stdout_hash": "e74c9a243ad50629ba564d0cec32b9e7e7ec0daf771f3ea30408ab82",
+    "stderr": null,
+    "stderr_hash": null,
+    "returncode": 0
+}
