+++ tests/tests.toml
@@ -294,7 +294,11 @@ filename = "../integration_tests/test_builtin.py"
 asr = true
 
 [[test]]
-filename = "../integration_tests/test_complex.py"
+filename = "../integration_tests/test_complex_01.py"
+asr = true
+
+[[test]]
+filename = "../integration_tests/test_complex_02.py"
 asr = true
 
 [[test]]
